Created CaptureAppConfig type
